requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.103.1
uvicorn==0.27.1
pillow==9.4.0
torch==1.13.1
tqdm==4.64.1
numpy==1.23.5
git+https://github.com/openai/CLIP.git
scikit-learn==1.1.3
joblib==1.2.0
pyyaml==6.0
qdrant-client==1.7.3
ftfy==6.1.3
